api(v1beta1): pre-freeze cleanup toward GA — dead fields, caBundleSecretRef, int32
Prepare the v1beta1 API for the v0.4.0 GA freeze (per the GA-API-FREEZE review):
- Remove dead Beskar7Machine.Spec.ConfigurationURL (breaking; consumed nowhere).
- Normalize RedfishConnection.CABundleSecretRef *LocalObjectReference -> string,
  matching the sibling CredentialsSecretRef (breaking); mutual-exclusion with
  InsecureSkipVerify preserved (now checks != "").
- HardwareRequirements (minCPUCores/minMemoryGB/minDiskGB) + inspection ints
  (cores/threads/sizeGB) int -> int32 (Kubernetes API convention).
- Delete dead RedfishConnectionInfo type.
- Fix InspectionImageURL godoc: it is the base URL under which vmlinuz/initrd.img
  are served (boot_handler.go), not an iPXE boot script.

Regenerated CRDs + deepcopy + chart CRDs (make manifests + sync-chart-crds,
verified idempotent). Updated examples, the touched docs fields, and the affected
controllers/tests. make test + -race + golangci-lint clean; RBAC drift guard green.

+++ controllers/inspection_handler_test.go
@@ -216,7 +216,7 @@ var _ = Describe("Inspection HTTP handler (PR-5.2)", func() {
 			g.Expect(report.Model).To(Equal("Test-7000"))
 			g.Expect(report.SerialNumber).To(Equal("SN-12345"))
 			g.Expect(report.CPUs).To(HaveLen(1))
-			g.Expect(report.CPUs[0].Cores).To(Equal(8))
+			g.Expect(report.CPUs[0].Cores).To(Equal(int32(8)))
 			// Owner reference back to PhysicalHost so the CM is GC'd on host delete.
 			g.Expect(cm.OwnerReferences).NotTo(BeEmpty())
 			g.Expect(cm.OwnerReferences[0].Kind).To(Equal("PhysicalHost"))
